+++ src/main/java/org/springframework/beans/factory/BeanFactory.java
@@ -31,4 +31,6 @@ public interface BeanFactory {
 	<T> T getBean(String name, Class<T> requiredType) throws BeansException;
 
 	<T> T getBean(Class<T> requiredType) throws BeansException;
+
+	boolean containsBean(String name);
 }