+++ src/main/java/org/springframework/beans/factory/annotation/AutowiredAnnotationBeanPostProcessor.java
@@ -1,12 +1,14 @@
 package org.springframework.beans.factory.annotation;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.util.TypeUtil;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.PropertyValues;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
 import org.springframework.beans.factory.ConfigurableListableBeanFactory;
 import org.springframework.beans.factory.config.InstantiationAwareBeanPostProcessor;
+import org.springframework.core.convert.ConversionService;
 
 import java.lang.reflect.Field;
 
@@ -33,8 +35,19 @@ public class AutowiredAnnotationBeanPostProcessor implements InstantiationAwareB
 		for (Field field : fields) {
 			Value valueAnnotation = field.getAnnotation(Value.class);
 			if (valueAnnotation != null) {
-				String value = valueAnnotation.value();
-				value = beanFactory.resolveEmbeddedValue(value);
+				Object value = valueAnnotation.value();
+				value = beanFactory.resolveEmbeddedValue((String) value);
+
+				//类型转换
+				Class<?> sourceType = value.getClass();
+				Class<?> targetType = (Class<?>) TypeUtil.getType(field);
+				ConversionService conversionService = beanFactory.getConversionService();
+				if (conversionService != null) {
+					if (conversionService.canConvert(sourceType, targetType)) {
+						value = conversionService.convert(value, targetType);
+					}
+				}
+
 				BeanUtil.setFieldValue(bean, field.getName(), value);
 			}
 		}