+++ src/main/java/org/springframework/beans/factory/config/ConfigurableBeanFactory.java
@@ -1,6 +1,7 @@
 package org.springframework.beans.factory.config;
 
 import org.springframework.beans.factory.HierarchicalBeanFactory;
+import org.springframework.core.convert.ConversionService;
 import org.springframework.util.StringValueResolver;
 
 /**
@@ -22,4 +23,9 @@ public interface ConfigurableBeanFactory extends HierarchicalBeanFactory, Single
 	void addEmbeddedValueResolver(StringValueResolver valueResolver);
 
 	String resolveEmbeddedValue(String value);
+
+	void setConversionService(ConversionService conversionService);
+
+	ConversionService getConversionService();
+
 }