+++ src/main/java/org/springframework/beans/factory/support/AbstractBeanFactory.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.FactoryBean;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.core.convert.ConversionService;
 import org.springframework.util.StringValueResolver;
 
 import java.util.ArrayList;
@@ -23,6 +24,9 @@ public abstract class AbstractBeanFactory extends DefaultSingletonBeanRegistry i
 	private final Map<String, Object> factoryBeanObjectCache = new HashMap<>();
 
 	private final List<StringValueResolver> embeddedValueResolvers = new ArrayList<StringValueResolver>();
+
+	private ConversionService conversionService;
+
 
 	@Override
 	public Object getBean(String name) throws BeansException {
@@ -73,6 +77,13 @@ public abstract class AbstractBeanFactory extends DefaultSingletonBeanRegistry i
 		return ((T) getBean(name));
 	}
 
+	@Override
+	public boolean containsBean(String name) {
+		return containsBeanDefinition(name);
+	}
+
+	protected abstract boolean containsBeanDefinition(String beanName);
+
 	protected abstract Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException;
 
 	protected abstract BeanDefinition getBeanDefinition(String beanName) throws BeansException;
@@ -98,5 +109,15 @@ public abstract class AbstractBeanFactory extends DefaultSingletonBeanRegistry i
 			result = resolver.resolveStringValue(result);
 		}
 		return result;
+	}
+
+	@Override
+	public ConversionService getConversionService() {
+		return conversionService;
+	}
+
+	@Override
+	public void setConversionService(ConversionService conversionService) {
+		this.conversionService = conversionService;
 	}
 }