+++ src/main/java/org/springframework/context/support/AbstractApplicationContext.java
@@ -11,6 +11,7 @@ import org.springframework.context.event.ApplicationEventMulticaster;
 import org.springframework.context.event.ContextClosedEvent;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.context.event.SimpleApplicationEventMulticaster;
+import org.springframework.core.convert.ConversionService;
 import org.springframework.core.io.DefaultResourceLoader;
 
 import java.util.Collection;
@@ -25,6 +26,8 @@ import java.util.Map;
 public abstract class AbstractApplicationContext extends DefaultResourceLoader implements ConfigurableApplicationContext {
 
 	public static final String APPLICATION_EVENT_MULTICASTER_BEAN_NAME = "applicationEventMulticaster";
+
+	public static final String CONVERSION_SERVICE_BEAN_NAME = "conversionService";
 
 	private ApplicationEventMulticaster applicationEventMulticaster;
 
@@ -49,11 +52,24 @@ public abstract class AbstractApplicationContext extends DefaultResourceLoader i
 		//注册事件监听器
 		registerListeners();
 
-		//提前实例化单例bean
-		beanFactory.preInstantiateSingletons();
+		//注册类型转换器和提前实例化单例bean
+		finishBeanFactoryInitialization(beanFactory);
 
 		//发布容器刷新完成事件
 		finishRefresh();
+	}
+
+	protected void finishBeanFactoryInitialization(ConfigurableListableBeanFactory beanFactory) {
+		//设置类型转换器
+		if (beanFactory.containsBean(CONVERSION_SERVICE_BEAN_NAME)) {
+			Object conversionService = beanFactory.getBean(CONVERSION_SERVICE_BEAN_NAME);
+			if (conversionService instanceof ConversionService) {
+				beanFactory.setConversionService((ConversionService) conversionService);
+			}
+		}
+
+		//提前实例化单例bean
+		beanFactory.preInstantiateSingletons();
 	}
 
 	/**
@@ -116,6 +132,11 @@ public abstract class AbstractApplicationContext extends DefaultResourceLoader i
 	@Override
 	public void publishEvent(ApplicationEvent event) {
 		applicationEventMulticaster.multicastEvent(event);
+	}
+
+	@Override
+	public boolean containsBean(String name) {
+		return getBeanFactory().containsBean(name);
 	}
 
 	@Override