+++ src/main/java/org/springframework/core/convert/support/GenericConversionService.java
@@ -1,5 +1,6 @@
 package org.springframework.core.convert.support;
 
+import cn.hutool.core.convert.BasicType;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.core.convert.converter.ConverterFactory;
@@ -28,6 +29,7 @@ public class GenericConversionService implements ConversionService, ConverterReg
 	@Override
 	public <T> T convert(Object source, Class<T> targetType) {
 		Class<?> sourceType = source.getClass();
+		targetType = (Class<T>) BasicType.wrap(targetType);
 		GenericConverter converter = getConverter(sourceType, targetType);
 		return (T) converter.convert(source, sourceType, targetType);
 	}
@@ -83,6 +85,8 @@ public class GenericConversionService implements ConversionService, ConverterReg
 
 	private List<Class<?>> getClassHierarchy(Class<?> clazz) {
 		List<Class<?>> hierarchy = new ArrayList<>();
+		//原始类转为包装类
+		clazz = BasicType.wrap(clazz);
 		while (clazz != null) {
 			hierarchy.add(clazz);
 			clazz = clazz.getSuperclass();