+++ src/test/java/org/springframework/test/common/StringToLocalDateConverter.java
@@ -1,0 +1,24 @@
+package org.springframework.test.common;
+
+import org.springframework.core.convert.converter.Converter;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * @author derekyi
+ * @date 2021/1/17
+ */
+public class StringToLocalDateConverter implements Converter<String, LocalDate> {
+
+	private final DateTimeFormatter DATE_TIME_FORMATTER;
+
+	public StringToLocalDateConverter(String pattern) {
+		DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern(pattern);
+	}
+
+	@Override
+	public LocalDate convert(String source) {
+		return LocalDate.parse(source, DATE_TIME_FORMATTER);
+	}
+}