+++ src/test/java/org/springframework/test/ioc/TypeConversionSecondPartTest.java
@@ -1,0 +1,25 @@
+package org.springframework.test.ioc;
+
+import org.junit.Test;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+import org.springframework.test.bean.Car;
+
+import java.time.LocalDate;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * @author derekyi
+ * @date 2021/1/17
+ */
+public class TypeConversionSecondPartTest {
+
+	@Test
+	public void testConversionService() throws Exception {
+		ClassPathXmlApplicationContext applicationContext = new ClassPathXmlApplicationContext("classpath:type-conversion-second-part.xml");
+
+		Car car = applicationContext.getBean("car", Car.class);
+		assertThat(car.getPrice()).isEqualTo(1000000);
+		assertThat(car.getProduceDate()).isEqualTo(LocalDate.of(2021, 1, 1));
+	}
+}